changed deletion of users
- a user can delete itself using their latest acces-token
- an admin can delete single users with their userID (ONLY for debugging purposes)

+++ src/main/java/at/termftp/backend/api/UserController.java
@@ -28,6 +28,8 @@ public class UserController {
     private final ConfirmationTokenService confirmationTokenService;
     private final EmailSenderService emailSenderService;
 
+    private final String TEMP_SECRET_ROOT_AT = "6293eed0e1as5e0413eed0e1klzasdffdedfR00!entasdfrph0";
+
     @Autowired
     public UserController(UserService userService, AccessTokenService accessTokenService, ConfirmationTokenService confirmationTokenService, EmailSenderService emailSenderService) {
         this.userService = userService;
@@ -85,18 +87,46 @@ public class UserController {
 
     /**
      * used to delete a single user by an ID
-     * TODO: check accesstoken
      * @param userID id of the User
+     * @apiNote should be only used for debugging
      * @return number of deleted users (int) or an error (Error)
      */
     @DeleteMapping(path = "/deleteUser/{id}")
-    public Object deleteUser(@PathVariable("id") String userID){
+    public Object deleteUserAsAdmin(@PathVariable("id") String userID,
+                                    @RequestHeader("Access-Token") String token){
+        if(!token.equals(TEMP_SECRET_ROOT_AT)){
+            return ResponseEntity.status(401)
+                    .body(new DefaultResponse(401, "Unauthorized", "Invalid Access-Token"));
+        }
+
         UUID id;
         try{
             id = UUID.fromString(userID);
             User user = userService.getUserById(id);
             int deletedUsers = userService.deleteUser(user);
             return DefaultResponse.createResponse(deletedUsers, "Deleted Users");
+        }catch (IllegalArgumentException ex){
+            return ResponseEntity.status(400).body(new DefaultResponse(400, "Bad Request", ErrorMessages.getInvalidUserID()));
+        }
+    }
+
+    /**
+     * used to delete a single user
+     * @param token the user's access token
+     * @return defaultResponse with status
+     */
+    @DeleteMapping(path = "/deleteUser")
+    public Object deleteUser(@RequestHeader("Access-Token") String token){
+        User user = accessTokenService.getUserByAccessToken(token);
+        if(user == null){
+            return ResponseEntity.status(401)
+                    .body(new DefaultResponse(401, "Unauthorized", "Invalid Access-Token"));
+        }
+
+        try{
+            int deletedUsers = userService.deleteUser(user);
+            System.out.println("- Deleted User " + user.getUsername());
+            return DefaultResponse.createResponse(deletedUsers, "Deleted User");
         }catch (IllegalArgumentException ex){
             return ResponseEntity.status(400).body(new DefaultResponse(400, "Bad Request", ErrorMessages.getInvalidUserID()));
         }
